requirements/requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    ./compile_requirements.sh
#
certifi==2019.11.28       # via http3, requests
chardet==3.0.4            # via http3, requests
click==7.0                # via uvicorn
fastapi==0.47.1
h11==0.8.1                # via http3, uvicorn
h2==3.1.1                 # via http3
hpack==3.0.0              # via h2
http3==0.6.7              # via requests-async
httptools==0.0.13         # via uvicorn
hyperframe==5.2.0         # via h2
idna==2.8                 # via http3, requests
pydantic==1.4             # via fastapi
python-dateutil==2.8.1
requests-async==0.6.2
requests==2.22.0
rfc3986==1.3.2            # via http3
six==1.14.0               # via python-dateutil
starlette==0.12.9         # via fastapi
urllib3==1.25.8           # via requests
uvicorn==0.11.2
uvloop==0.14.0            # via uvicorn
websockets==8.1           # via uvicorn
